requirements.txt
jupyterlab
pyperclip
pyyaml
Send2Trash